requirements.txt
pypdf==5.0.0
pikepdf==9.5.2
PyMuPDF==1.24.10
pdfminer.six==20231228
pdfplumber==0.11.4
reportlab==4.2.5
pdf2image==1.17.0
Pillow==10.4.0
img2pdf==0.5.1
pytesseract==0.3.13
ocrmypdf==16.9.0
typer==0.12.5
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
